Modified message writing behavior and keyboard interaction on msgs pane.
When first entering the messages, msg textbox is not focused. You need
to type 'm' or put the cursor with the mouse to start a message. Once you
write a message, the msg textbox stays focused until you click out of
the message textbox or hit 'enter' with an empty message.