Enabled base mechanism for tabs & notifications configuration.
Tabs: by default if convore is already opened in a tab an a user clicks a link
in the chrome extension that opens in convore, it will use the last convore
opened tab to display the new page.
Other thought option will be 'always open in new tab'.

Notifications: by default they always show unless the convore web page
is the focused active tab in which case they won't show to inhibit
double notification (as Convore's web app already does notify).
Other thought options are: 'always' (notified even if the convore page
is the focused tab) and 'never'.